refactor: dto 패키지 추가 및 클래스 이동
