+++ src/main/java/roomescape/controller/ReservationController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
-import roomescape.ReservationDto;
+import roomescape.dto.ReservationDto;
 import roomescape.domain.ReservationRepository;
 
 @RestController
