+++ src/test/java/roomescape/ReservationTestSetting.java
@@ -4,6 +4,7 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Objects;
 import roomescape.domain.Reservation;
+import roomescape.dto.ReservationDto;
 
 public class ReservationTestSetting {
